wiki_api/requirements.txt
flask
Flask-Cors
flask-swagger
python-dotenv
requests
